backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
pytesseract==0.3.10
openai==1.3.7
reportlab==4.0.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
elevenlabs==0.2.27
whisper==1.1.10
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
